Update ChessPanel.java
This doesn't change any of the functionality of the code. All the new code is commented out because it doesn't work, I just want some help on it. I'm trying to get it to display a message if the user is in checkmate, but it keeps crashing whenever I get the game into a checkmate. (By the way, in order to get it to a checkmate quickly for testing, look up how to do a Fool's Mate, which ends the game in two turns.)